room and meassage function 3:27/7

+++ base/views.py
@@ -1,8 +1,61 @@
+from django.http.response import HttpResponse
 from django.shortcuts import render, redirect
-from . models import Room, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
+from . models import Message, Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
 # Create your views here.
+
+
+def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+
+        except:
+            messages.error(request, 'User dose not excist')
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Usernaem or Passwrod not exict')
+
+    contxt = {
+        'page': 'login'
+    }
+    return render(request, 'base/login_register.html', contxt)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registeration')
+    contxt = {
+        'form': form,
+    }
+    return render(request, 'base/login_register.html', contxt)
 
 
 def home(request):
@@ -24,12 +77,26 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all().order_by('-created')
+    participants = room.participants.all()
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk=room.id)
     contxt = {
-        "room": room
+        "room": room,
+        "room_messages": room_messages,
+        'participants': participants
+
     }
     return render(request, "base/room.html", contxt)
 
 
+@login_required(login_url='login')
 def createroom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -44,9 +111,13 @@ def createroom(request):
     return render(request, 'base/room_form.html', contxt)
 
 
+@login_required(login_url='login')
 def updateroom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -58,12 +129,29 @@ def updateroom(request, pk):
     return render(request, 'base/room_form.html', contxt)
 
 
+@login_required(login_url='login')
 def deleteroom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     contxt = {
         "obj": room
+    }
+    return render(request, 'base/delete.html', contxt)
+
+
+@login_required(login_url='login')
+def deletemessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!!')
+    if request.method == 'POST':
+        message.delete()
+        return redirect('room', pk=message.room.id)
+    contxt = {
+        "obj": message
     }
     return render(request, 'base/delete.html', contxt)
